Completed add product step

+++ bot/handlers/seller/helpers.py
@@ -15,17 +15,6 @@ from bot.utils.message_utils.message_utils import MessageSetting, send_message, 
 from bot.utils.message_utils.media_messages_utils import send_media_message, send_cached_media_message
 
 
-async def add_field_product(bot: Bot, state: FSMContext, config: FieldConfig) -> tuple[MessageSetting | None, State]:
-    if not config.is_end_field:
-        next_config: FieldConfig = ADD_FIELD_PRODUCT_CONFIGS[config.next_field]
-        return next_config.input_msg, config.next_state
-    else:
-        return await complete_product(state, bot)
-
-async def edit_field_product(bot: Bot, state: FSMContext) -> tuple[MessageSetting | None, State]:
-    return await complete_product(state, bot)
-
-
 async def handler_input_product_field(msg: Message, state: FSMContext, field_name: str, value,
                                       is_delete_user_message: bool=True):
     if is_delete_user_message:
@@ -44,10 +33,11 @@ async def handler_input_product_field(msg: Message, state: FSMContext, field_nam
         await state.update_data(**{ParamFSM.SellerData.ADD_PRODUCT_OPERATOR: product})
         now_state: str = await state.get_state()
         next_state: State | None = None
-        if now_state.startswith(AddProductStates.group_name):
-            new_message, next_state = await add_field_product(msg.bot, state, filed_config)
-        elif now_state.startswith(EditProductStates.group_name):
-            new_message, next_state = await edit_field_product(msg.bot, state)
+        if now_state.startswith(EditProductStates.group_name) or filed_config.is_end_field:
+            new_message, next_state = await complete_product(state, msg.bot)
+        elif now_state.startswith(AddProductStates.group_name):
+            next_config: FieldConfig = ADD_FIELD_PRODUCT_CONFIGS[filed_config.next_field]
+            new_message, next_state = next_config.input_msg, filed_config.next_state
         else:
             ValueError(f'Wrong state: {now_state}')
         await state.set_state(next_state)
@@ -59,18 +49,24 @@ async def handler_input_product_field(msg: Message, state: FSMContext, field_nam
 
 
 async def complete_product(state: FSMContext, bot: Bot) -> tuple[None, State]:
-    await delete_bot_message(state, bot)
-
     product: InputProduct
     (product, chat_id) = await get_data_state(state, ParamFSM.SellerData.ADD_PRODUCT_OPERATOR,
                                                        ParamFSM.BotMessagesData.CHAT_ID)
-    print(product.media)
-    product_message = MessageSetting(text=ADD_PRODUCT_FORM_TEXT.insert((product.name,
-                                                                          product.catalog,
-                                                                          product.description,
-                                                                          product.price)),
-                                     cache_media=product.media.get_cache())
-    await send_cached_media_message(state, bot, product_message)
-    await send_message(state, bot, COMPLETE_ADD_PRODUCT_MESSAGE)
 
-    return None, AddProductStates.user_checking
+    new_message = None
+    product_data = (product.name,
+                    product.catalog,
+                    product.description,
+                    product.price)
+    complete_text = ADD_PRODUCT_FORM_TEXT.insert(product_data)
+    if product.media is not None:
+        await delete_bot_message(state, bot)
+        product_message = MessageSetting(text=complete_text,
+                                         cache_media=product.media.get_cache())
+        await send_cached_media_message(state, bot, product_message)
+        await send_message(state, bot, COMPLETE_ADD_PRODUCT_MESSAGE_WITH_MEDIA)
+    else:
+        new_message = MessageSetting(text=COMPLETE_ADD_PRODUCT_MESSAGE.insert((complete_text,)),
+                                     keyboard=ADD_PRODUCT_COMPLETE_KEYBOARD)
+
+    return new_message, AddProductStates.user_checking
